Fix GPU backend
CPU is now used as default device for all JITed functions.
GPU can be specified as a command line flag, along with an optional
device ID to use a particular device, otherwise the backend will
be queried and the device with the most available memory will be used.
GPU will only be used in continuation method for objective and
jacobian functions.