Use grad for scalar minimization instead of jacfwd
jax.grad uses reverse mode, which is waaaaay faster for the scalar
objective functions than forward mode which we were using.
(forward mode is still faster for the full newton method, and is still
used for that).
Also fixed calling signature in numpy version of jacfwd so it should
work correctly with either backend.