Fix bug with 'accel' method, change scalar obj function
The 'accel' method objective function was still using the 'rcond'
parameter when calling ZernikeTransform.fit(), which was removed
a while ago in favor of precomputing the pseudo-inverse.

Also changed scalar objectives to minimize log(1+f^2) rather than
just f^2. This seems to help keep the function and gradient magnitudes
more in line and helps convergence a bit.