Update to support JAX 0.2.5
-behavior of jnp.where changed, it now only allows the full usage,
where "true" and "false" values are given, so the output is the same
shape as the input.
-Refactored BC, objective, and backend functions to handle this
correctly, making sure we always specify the correct outputs.
-Also fixed some indexing issues in BC that gave the wrong answer
for nonsymmetric node patterns
-New version of JAX also complains when using numpy arrays as
static arguments to jitted functions, so now we just wrap them
in tuples before passing to jitted function.
-NaN gradients in reverse mode still an issue, if we want to fix it
we'll likely need to redo a lot of the field component calculations
to handle the axis better (instead of computing nans and then
overwriting, we need to avoid computing them in the first place).

Resolves #44