started changing Configuration to polar coordinates framework
Note: changed the input file format required by input_reader.py. This was not necessary for the polar coordinates, but I think it is an improvement.
Naming convention for spectral coefficients is now X_yz where X is the variable being represented and yz are the spectral indicies (a combo of l, m, and n).
Started re-writing the Configuration methods to handle the polar coordiantes; more work is needed.