refactored compute functions
Re-wrote all compute functions to take the same call signatures.
Compute functions now handle all calls to other compute functions internally, so the user does not need to know the dependencies.
Terms as computed as-needed for the final quantity.
Note: these functions now return a list of dicts, not just the requested dictionary.