Remove continuation.py
-All of continuation method now implemented as methods of EquilibraFamily
-input reader now returns a list of dictionaries rather than a
dictionary of lists. EquilibriaFamily can be initialized with a
list of dicts, and will then initialize an entire family of Equilibrium
at once
-These equilibria are then solved in sequence, using the previous solution
as an initial guess
-Also supports starting the solution at an arbitrary point, allowing for
resuming of previous attempts etc.
-using vmec as an initial guess now done in main by just setting the
first equilibrium to be the vmec one
-new utils for copying and expanding coefficient vectors

Still to do:
-need to implement perturbations
-each equilibria has its own transforms, which are recomputed each
time. might be better to copy the old ones to avoid recomputation.
also the possible issue of lots of memory usage if there are
many equilibria with direct transforms