Update perturbations with trust region
-Method of perturbations implicitly assumes an asymptotic ordering
of the terms in the series, but sometimes the 2nd order term would
be much larger than the first order and the result would be super wrong.
-Perturbations are now done using a trust region approach, where the
error is minimized subject to a bound on the step size, and the bound
is inversely proportional to the order of the perturbation.
-trust region ratio can be varied, default of 0.1 seems ok.
-2nd order perturbations for BC seem to work fine now
-2nd order for pressure still works, though visually they look a bit
worse despite the new method resulting in lower force error.