Update configuration profile resolution
-Using the same radial resolution for the equilibrium and the profiles
made sense initially, but it means that if the profiles require a lot of
modes, they get truncated for low res equilibria.
-Usually this isnt an issue but for some equilibria (eg NCSX) truncating
the power series for the profiles causes them to be suuuuper wrong
since all the coefficients are huge but they cancel nicely.
-Configuration now uses the max of the profile resolution or its own
radial resolution for the profile basis, so profiles will always be
full resolution, and will have extra zeros tacked on if the equilibrium
resolution is higher