Updates to fix IO and jit compilation
This commmit addresses 2 major issues: objectives/optimizers not being
saved, and objectives getting compiled more often than necessary

Changes to Equilibium/EquilibriaFamily:
-general switching to using properties rather than direct attributes
when referencing things (ie, eq.foo, not eq._foo). This allows getter
methods to have safeguards if things weren't defined or loaded correctly
for some reason
-Add "node_pattern" property to equilibrium
-Add public "transforms" property to equilibrium (public interface to
old _transforms dict)
-When assigning objective function to equilibrium, it now checks if the
new one is equivalent to the old one, if they are it skips the update.
This prevents needless recompilation if nothing really changed.
-optimizer and objective attributes now assigned to equilibrium.initial

Changes to objective functions:
-object lib is now set correctly for saving/loading
-init method can now properly handle loading from file
-moved most of the derivative setup/jit/etc to its own method that is
automatically called after the main init. The function "set_derivatives"
can also be called manually to change jit settings or devices to
compile to.
-compiling is now done on the objective rather than the optimizer,
again, a way to prevent needless recompilation. This is done with a new
"compile" method that takes the generic function arguments to call the
objective, plus a "mode" argument to tell it which derivatives to
compile (ie, for scalar vs least squares optimization)
-new "eq" method for comparing different objective functions.
Effectively the same way we've been doing a custom __eq__, but we can't
do that for the objectives because it breaks the hashing the jax uses
when jitting the objective. So instead of doing objective1 == objective2,
do objective1.eq(objective2)
-Removed init methods from ForceErrorNodes and ForceConstraintNodes,
since the default one from ObjectiveFunction now handles everything.
-Init for Galerkin and Energy remains but just calls super init and then
warns if the grid is not quadrature grid
-new method to make sure the transforms have the correct derivatives
for the objective and recomputing them if not

Changes to optimizer:
-io attributes now set, inheritance from IOAble and refactored init to
work with io stuff
-objective no longer passed in at init, just the method
-instead, objective is now passed as an argument to optimizer.optimize()
-removed compile method in favor of compiling the objective directly
(which is automatically done in optimizer.optimize)
-added equality checking for optimizers