Refactor zernike evaluation
- This commit addresses two major things: speed, and differentiability
- Previously, obtaining accurate zernike evaluation required using
extended precision arithmetic that meant we couldnt' use JAX for
that part of the calculation, making it both slow when the grid changes
on each iteration (as with computing inverse coordinate mappings and
field line tracing etc), and also non differentiable.
- A new method has been implemented that uses identities for the zernike
polynomials in terms of jacobi polynomials, and a stable iterative
evaluation for the jacobi polynomials and binomial coefficients.
- Accuracy seems on par or better than the old method using extended
precision, at least for a given amount of computation time.
- There is some overhead from JIT compilation, but seems to pay off well
for high resolution
- Also implemented a new "unique" option for basis.evaluate to first
reduce the work by finding unique combos of nodes/modes. Previously
this was done inside each basis function evaluation, but doing it on the
outside should be more efficient and makes the underlying functions
differentiable.

Resolves #117