symmetric current profile bug fix
If the initial current profile is a vacuum then it gets set to an "even" profile by default. But then it has no odd modes, and the typical current profile from an input file is odd (even power series integral = odd power series). This caused the current coefficients to not be perturbed properly.