Fix all notebooks except tutorials/03, rewrite tutorials/00....
Title plus
- Black'ed notebooks
- fixed a bunch of typos
- Rewrote jupyter 'raw' cells in markdown because jupyter 'raw' cells
are not readable by github or desc api webpage.
- tutorials/03 will still crash; we need to regenerate the
qs_initial_guess.h5 file to fix this. I think this is the QHS input.
I have a solve going on stellar that
seems to have stalled on step 6. Will see in morning.
- all the other notebooks are fixed, if github complains it's probably because
I added an extra '../' to the path of the tutorials notebooks