Change notebook kernel back to default ipykernel
I switched the jupyter notebook kernel to my conda environment last commit.
Pretty normal thing to do for jupyter lab with a conda env.
Apparently that get stored in the notebook,
so I switched the notebook kernels back to the default ipkernel in this
commit.
If github still complains that it can't find a kernel, then
I don't know how to fix that.