Add function_of key to data_index
If we want, we can change the key name / type using this terminal command
which is essentially a multiple file spanning find and replace command.

find . -type f -execdir sed --in-place 's/${regexp}/${replacement}/g' '{}' '+'

where the ${regexp} should be changed to the string to replace.
    for example: function_of="rtz"
and ${replacement} should be changed to the new string.
    for example: func_of=["r", "t", "z"]

Though a backslash will need to precede special characters such as forward
slashes.