Add second hyperparameter to free boundary objective
so that the local singular integration region can be
modeled as a circle in theta, zeta space or ideally as
a circle in real space by some clever choice from the user.

Apply the strategy of #1530 to bounce integral interpolation.
This is a crutch to hold out until the nuffts are added to DESC.

Move Fourier conversion utilities to vmec_utils as mentioned in #1531