anisotropy (#1546)
- [x] Generalizes the change of variable in the singular kernel
integrals to improves convergence of the integration. Before the
integration region was an NFP elongated ellipse which will slow down
convergence. That is changed here so that the region can be chosen to be
a circle in real space to account for the affect of the the aspect ratio
of the device has on the anisotropy of the grid.
- [x] Change default chunk size for `virtual_casing_biot_savart` and
free boundary

---------

Co-authored-by: Rory Conlin <roryconlin@gmail.com>
Co-authored-by: Dario Panici <dpanici@princeton.edu>