Dump Python 3.9, add support for 3.13 (#1621)
Before updating our dependencies, we should dump the Python 3.9 (and
optionally add 3.13). The latest JAX version supported by Python 3.9 is
0.4.30 and those unit, regression and benchmarks don't use the latest
JAX version, so we might miss some things.

Misc

- Adds a flag to skip the CI tests called `only-docs-comments`, if this
label is present none, of the tests will run.
- Removes `requirement` changes from benchmarks (so if the only change
is requirements file, benchmarks will not run). I changed this because,
in benchmarks we use the same environment (the requirement file from
feature branch) for both tests, so this is redundant. Optionally, we
might create new environments for both tests, let me know if you want
that to be the case.

---------

Co-authored-by: Kaya Unalmis <kayaunalmis@proton.me>