Make basis modes as static attribute (#1582)
According to the explanation `_AutoRegisterPytree` metaclass, basis
modes should be static and stored in `aux_data` but if you run,
```python
eq = desc.examples.get("ATF")
child, aux_data = tree_flatten(eq)
child
```
you can see that modes are stored in children.

This PR adds the basis arguments of some classes to `_static_attrs` such
that they are not traced during `jit` decorated functions. One important
point to consider when making variables static is that that variable
shouldn't be used in the computations. For example, we store `R_basis`
in `Equilibrium` class but we never use it directly inside compute
functions. The basis related stuff are stored in `constants`. That being
said, if we set `_modes` of `_Basis` class as static, we will get an
error with the latest changes of #1405 since full set of modes is sliced
by unique indices.

During testing, I have seen that this doesn't affect memory usage, and
in contrast to my expectations, it doesn't reduce the compilation time
either. The speed up mainly comes from the `tree_flatten` operation. So,
very simple compute functions (like ForceBalance) gets a major speed up.
But others not much.

---------

Co-authored-by: Kaya Unalmis <kayaunalmis@proton.me>