Optional memory profiler workflow (#1702)
This adds an optional profiling workflow which will profile the CPU
memory usage of some memory-intensive routines (such as Free Boundary,
Bounce Averages, Jacobian) and comment the results in a way similar to
timing benchmarks. Also, this will plot the data for each function, as
well as master plots and saves it to artifacts.

There are a couple of concerns. First, how reliable is this, which I
think is good enough to see significant changes. Second, I just run them
around 3 times, I think this is enough since memory profile is less
volatile then timing.

---------

Co-authored-by: Daniel Dudt <33005725+ddudt@users.noreply.github.com>