Initial Guess for Non-Convex Shapes (#1643)
For non-convex cross-sections, our initial guess function cannot form a
set of nested surfaces. This PR changes the axis parser to compute the
midpoint between theta=0 and theta=pi, and create an `FourierRZCurve`
from those.

Ideally, one could add the Shafranov shift with some heuristic on the
pressure profile but for now, I won't add since for Stellarators, the
shift is not as straightforward as tokamaks.

Resolves #1384

---------

Co-authored-by: Dario Panici <37969854+dpanici@users.noreply.github.com>
Co-authored-by: Rahul Gaur <19224702+rahulgaur104@users.noreply.github.com>