Add `blocked` option to `ProximalProjection` (#1623)
This separates the force balance part and objective part of
`ProximalProjection` `jac` methods and applies batch/blocked to both of
them. Basically, moves the `batched_vectorized` to objective part (but
optionally, where previously we were always batching it).

Additionally, instead of taking the full jacobian and then applying the
matrix vector multiplication (in this case this is exactly equivalent to
jvp), replaces it with jvps in proper directions. This reduces the
compute time for CPU (not for GPU).

This will be useful for concurrently calculating the jacobian of each
objective in #1495

- Fixes the bug where `blocked` mode of proximal was not using each
sub-objectives `jac_chunk_size`
- Adds a lot of code explanations (especially for variable dimensions)
- Renames `_unfixed_idx_mat` to `_feasible_tangents`
- Adds `_eq_` prefix to some variable names to make it clear that they
belong to the `Equilibrium`
 
Resolves #1589

---------

Co-authored-by: Dario Panici <37969854+dpanici@users.noreply.github.com>
Co-authored-by: Dario Panici <dpanici@princeton.edu>