Add PlasmaCoilSetMaxDistance objective (#1722)
Sometimes, the coils get too far away from the plasma, and I often need
an objective function which constrains the distance between plasma and
coil from above. This adds that objective by generalizing the
PlasmaCoilSetDistance objective to find the maximum (over the coil)
minimum (over the plasma) distance between coil and plasma. To do this,
I also added functionality to the `softmax` function to allow it to
perform maximization over a given axis rather than flattening the array.

---------

Co-authored-by: Dario Panici <37969854+dpanici@users.noreply.github.com>
Co-authored-by: Yigit Gunsur Elmacioglu <102380275+YigitElma@users.noreply.github.com>