Integrate on boundary for more quantities. Enforce use of full grid when required. (#1094)
- [x] Some quantities like the plasma volume `V`, major radius `R0`,
minor radius `a` etc. can be computed via stoke's theorem on the last
closed flux surface. This pull request modifies the compute functions to
default to such an integration on the boundary rather than throughout
the volume so that the length scale quantities can be computed
accurately on grids other than `QuadratureGrid`. This lets us avoid
building transforms for `QuadratureGrid` in objectives, saving memory.
Also, has some improvements in general

> 1. Reducing dimensionality of integral makes integration more
tractable with faster convergence.
> 2. Transforming the integration to one over periodic domains makes it
more accurate due to the exponential convergence of the quadrature on
periodic domains.

- [x] Merging this PR completes my list of reviewing tasks to resolve
#1101

---------

Co-authored-by: Daniel Dudt <33005725+ddudt@users.noreply.github.com>
Co-authored-by: Dario Panici <37969854+dpanici@users.noreply.github.com>
Co-authored-by: Yigit Gunsur Elmacioglu <102380275+YigitElma@users.noreply.github.com>
Co-authored-by: Dario G. Panici <dpanici@princeton.edu>