Improve documentation for ballooning (#1782)
This pull request documents the `g ballooning`, `c ballooning`, and `f
ballooning` ballooning quantities to be the variables defined in
equations 25 in
[arxiv.org/abs/2410.04576](https://arxiv.org/abs/2410.04576) with the
following differences.

1. `g ballooning`, `c ballooning`, and `f ballooning` are the
dimeinsionless forms of equation 25.
2. `g ballooning`, `c ballooning`, and `f ballooning` have an additional
scale factor of ``rho**2`` compared to equation 25.
3. `c ballooning` has an additional factor `sign(ψ)` compared to
equation 25.
4. `g ballooning` and `f ballooning` redefine `α = α_{DESC} + ι ζ₀ sign
ι`
5. `c ballooning` redefines `α = α_{DESC} + ι ζ₀`
6. Also, the input to these functions in the code is labeled `zeta0` but
the expectation is for the user to pass in `ι ζ₀` (and not `ζ₀` and also
not `ι ζ₀ sign ι`).

This PR documents these things and corrects the description of the
formulas in comments on `master`. The actual computation is the same as
`master`.