don't close over field in field_line_integrate (#1830)
Don't close over `field` to avoid long compile times due to constant
folding when the field has large array attributes.

This helps dramatically reduce compile times when constant folding would
be slow, and seems to have no effect on runtimes in cases where there
would not be appreciably slow constant folding occuring, so it seems to
be a harmless change.

- [x] check effect on GPU (before, running on gpu took a surprisingly
large amount of memory, even if it could run easily on the cpu with less
memory) - this PR also improves speed for GPU, have not profiled memory.

Resolves #1829