Partial summation in coordinate mapping (#1826)
Progress toward addressing #1154.

The [attached
benchmark](https://github.com/user-attachments/files/21745848/benchmark_partial_sum.zip)
must be ran on `ku/nufft` branch.
250x speed improvement from ~15 seconds to ~50 milliseconds.
The improvement would be more if the FourierZernike basis is padded as
discussed in #1243.
Then we would avoid the 3D spectral to real transform as well as N^2
FFTs of size N.
(Then this computation would likely be in microsecond range).