Use ScaledProfile in rescale (#1871)
Resolves #1870 

I chose to just use `ScaledProfile` because we have other profile types
where the scaling could be incorrect as well, such as in the case of
`PowerProfile` (where the power it is raised to would be scaled, and the
sub-profile coefficients would also be scaled, resulting in a completely
incorrect rescaling). Using `ScaledProfile` is the only way to get this
to work correctly for all profile types.

I could maybe see the case for adding an instance check to see if the
profiles are `PowerSeriesProfile` instances, and in those cases doing
the multiplication, in order to preserve the current behavior of it not
modifying power series profiles, but I would rather just use the
`ScaledProfile`.

---------

Co-authored-by: Yigit Gunsur Elmacioglu <102380275+YigitElma@users.noreply.github.com>