Prevent weak type promotion to reduce recompilation (#1963)
Resolves #1581 

Some of the attributes in optimizable classes are cast to float by
`float()` function which returns a Python value, but internally JAX
converts them to `jnp.float64` before jit, this causes recompilation in
some cases. Technically, the number of compilations shouldn't be more
than 2, but this can be the cause of some recompilations.

I know there has been some discussion on not using `jnp.float64` before
to leave some room for #1048, but I don't see that working in the near
term.

```python
jax.config.update('jax_explain_cache_misses', True)
```
This was very useful to pinpoint the exact cause of recompilation in
#1581

---------

Co-authored-by: Dario Panici <dpanici@princeton.edu>