Always ensure that equilibrium profiles are consistent with eq.L (#1982)
This is one way to resolve the issue in #1974, by just ensuring eq
profile resolution is at least or higher than the eq radial resolutions
always. This also should work fine in the cases of non-power-series
profiles, as those would be the same as far as number of params between
different eq resolutions.

Resolves #1974

---------

Co-authored-by: YigitElma <yigitelmacioglu@gmail.com>
Co-authored-by: Yigit Gunsur Elmacioglu <102380275+YigitElma@users.noreply.github.com>