Add `ShareParameters` Linear Objective (#1320)
If someone wants to help with any of below feel free, I just got
motivated for the initial implementation.


- Adds `ShareParameters` linear objective, which takes in a list of
arbitrary length of objects and a dict of `params` to fix which
  - [x] must all be the same type  (implemented check)
- [x] be the same resolutions in the params being shared (i.e. equal
`thing.dimensions[key]` where key is e.g. "R_lmn")

The objective then will fix the indices of the parameters of each object
to the same value as each other object.

TODO:
- [x] make sure implementation works for pytree objects such as
`CoilSet` (tests added)
- [x] add check for `params` lengths being the same across objects
(technically, we also should be checking that, say, if `p_l` is being
fixed btwn two equilibria, that they are also the same TYPE of pressure
profile, but not sure how we can go about doing that without having very
individualized logic like `_check_type` does in `coils.py`)
- [x] Add test for fixing only specific indices of a param
- [x] Add more test for coils (to test pytree objects, I am not sure of
all the use case right now and so can't think of any good tests)

Known issue:
- for attributes like `FourierRZCurve.normal` where we normalize the
vector when the attribute is set, if I for instance do
`ShareParameters([coil1, coil2], {"normal":True})` everything works
perfect, but if I do though `ShareParameters([coil1, coil2],
{"normal":np.array([1])})` to only make the index-1 component of the
normal the same, the constraint works correctly (as in if you `project`
and `recover`, that component will be the same), but when the attributes
are then set at the end of the optimization, they are normalized in the
setter, and the 0th and 2nd components may differ (as they were not
shared and thus can be different), and the resulting normalized vector
may have a different 1-component due to the potentially different
overall magnitudes of the un-normalized normal vectors

This is sort of a rare and kinda useless edge case, but still unsure of
how to fully resolve a case like this. It is hard to enforce that the
normalized vector component be the same as that is no longer a linear
constraint, and the normalization happens after the optimization. I
think we could maybe just put a disclaimer in the docstring for this
case maybe?

Resolves #1250

---------

Co-authored-by: Yigit Gunsur Elmacioglu <102380275+YigitElma@users.noreply.github.com>
Co-authored-by: YigitElma <yigitelmacioglu@gmail.com>
Co-authored-by: Rahul Gaur <19224702+rahulgaur104@users.noreply.github.com>